deleted useless files and merge types together
