+++ src/helper.ts
@@ -1,6 +1,6 @@
-import { API } from '.';
 import { SERVER_CONNECTOR_EXTENSION_ID } from './constants';
 import * as vscode from 'vscode';
+import { API } from './util/types';
 
 async function activateExtension(): Promise<API | undefined>  {
     const extension = vscode.extensions.getExtension<API>(SERVER_CONNECTOR_EXTENSION_ID);
@@ -11,7 +11,7 @@ async function activateExtension(): Promise<API | undefined>  {
     return api;
 }
 
-export async function getAPI(): Promise<API> {
+export async function retrieveUIExtension(): Promise<API> {
     const api = await activateExtension();
     if (!api) {
         return {
