+++ src/main.ts
@@ -1,5 +1,4 @@
 export * from './constants';
-export * from './index';
 export * from './api/rspModel';
 export * from './main';
 export * from './util/types';
